converted scannercontrol to sidebar for better UI experience
