+++ spynet/src/Sidebar.css
@@ -1,0 +1,110 @@
+/* Sidebar.css */
+
+.sidebar {
+    position: fixed;
+    top: 0;
+    left: 0;
+    height: 100%;
+    width: 300px; /* expanded width */
+    background-color: #1e1e1e;
+    border-right: 1px solid #333;
+    padding: 15px;
+    box-sizing: border-box;
+    transition: width 0.3s ease;
+    overflow: hidden;
+    z-index: 1000;
+  }
+  
+  .sidebar.collapsed {
+    width: 40px; /* collapsed width */
+  }
+  
+  .toggle-btn {
+    position: absolute;
+    top: 10px;
+    right: -25px;
+    background-color: #333;
+    border: none;
+    color: #fff;
+    width: 25px;
+    height: 25px;
+    cursor: pointer;
+    border-radius: 3px;
+    outline: none;
+  }
+  
+  .controls {
+    margin-top: 40px; /* space for toggle button */
+  }
+  
+  .form-row {
+    margin-bottom: 10px;
+    display: flex;
+    flex-direction: column;
+  }
+  
+  .form-row label {
+    margin-bottom: 5px;
+    color: #bbb;
+    font-size: 0.9em;
+  }
+  
+  .form-row input {
+    padding: 8px;
+    border: 1px solid #444;
+    border-radius: 4px;
+    background-color: #2a2a2a;
+    color: #fff;
+    font-size: 1em;
+    outline: none;
+  }
+  
+  .button-group {
+    display: flex;
+    flex-wrap: wrap;
+    gap: 10px;
+    margin-top: 15px;
+  }
+  
+  .button-group .btn {
+    flex: 1;
+    padding: 10px;
+    border: none;
+    border-radius: 4px;
+    cursor: pointer;
+    color: #fff;
+    transition: background-color 0.3s ease;
+  }
+  
+  .btn-start {
+    background-color: #4caf50;
+  }
+  
+  .btn-start:hover {
+    background-color: #45a049;
+  }
+  
+  .btn-pause {
+    background-color: #ff9800;
+  }
+  
+  .btn-pause:hover {
+    background-color: #fb8c00;
+  }
+  
+  .btn-resume {
+    background-color: #2196f3;
+  }
+  
+  .btn-resume:hover {
+    background-color: #1e88e5;
+  }
+  
+  .btn-stop {
+    background-color: #f44336;
+  }
+  
+  .btn-stop:hover {
+    background-color: #e53935;
+  }
+
